Drop the host network metrics that are reported for host pods in docker
If docker is used as container runtime, the dummy container name "POD"
is used by the kubelet/cadvisor for the container label.

The container network metrics are reported for each pod: both that are
in the host network namespace or in the pod's network namespace.

The metrics for the host network namespace are meaningless, they are
not specific to the pod. So we drop these metrics, except for the eth0
interface. In golang, we can not use negative lookaheads in regular
expressions, so we opted for the workaround to dropping interfaces that
are longer than 5 characters: like cali.+, cilium, lxc...

This is an optimization for existing docker based k8s clusters.
We observed that the network metrics can have a significant cardinality
in larger clusters. (2k valuable metrics out of 50k)

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>