Add a label host_network="true" for network metrics of the host
The id is the cgroup path.
- for pods, it is something like /kubepods/burstable/...
- for the host, the id is "/"

The host level network metrics can be differentiated from the pod level
network metrics based on this new label "host_network" in prometheus
queries.

Note that the id label itself is dropped by the next rule, which is fine
because we don't need the details of the id label.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>