Align the cadvisor related relabeling rules for the cache prometheus
The prometheus in the control plane scrapes the kubelet's cadvisor
metrics from the shoot cluster nodes. It is used to get metrics about
components that are deployed into the kube-system namespace of the shoot.

The "cache" prometheus in the garden namespace of the seed cluster scrapes
the kubelet's cadvisor metrics from the seed cluster nodes. It is used
to get metrics about the components that are running in the control plane,
in the seed. (The prometheus instances (up to 200 per seed) in the control
plane federate the relevant metrics from the "cache" prometheus and
thereby avoid scraping all the cadvisor metrics from the kubelet 200
times.)

The previous commits adjusted the cadvisor scrape configuration of the
control plane prometheus. This commit applies the same changes to the
cache prometheus. Please refer to those commits for a detailed explanation.

With this commit, the container_network_.+ metrics are going to be
kept properly for clusters that use containerd as container runtime.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>