Improve the Kubernetes Pods dashboard: no network metrics for host pods
The network I/O panel of the Kubernetes Pods dashboard didn't work
for clusters with conainerd. The previous commits fixed the issue
in the cadvisor scrape configuration that led to dropping the network
metrics for containerd clusters. With those changes in place,
the network metrics are shown as expected (no change was needed in the
dashboard).

During the investigation we also understood that previously network
metrics were shown for pods that are in the host network namespace.
It turns out that it is not really meaningful to show network metrics
for pods that run in the host network namespace. The metrics are not
specific to the pod, but to the host network namespace and hence contain
traffic from system services and all the other pods on the node.

Instead of showing the host network metrics for pods in the host network
namespace, we rather decided to not show anything. To avoid the confusion
caused by "missing data", we added a table left to the Network I/O panel
which shows if the selected pods are in the host network namespace and
explains in its description that for those pods, there are no network
metrics.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>