Fix the node level network metrics on the Node Details dashboard
Previously, the Network I/O pressure panel used a query which was
unfortunately incorrect and yielded bogus results.
It used to sum up all the network traffic from all the pods on a host.
It is incorrect because
- it does not consider inter pod traffic on the specific node,
  when 2 pods on the same node talk to each other and the traffic
  might not leave the node itself
- it does not consider that for pods in the host network namespace
  the host network devices are reported and adding up those metrics means
  that the host network traffic is counted several times

Based on the changes in the previous commits, this panel shows the
host network metrics of the host as it is exposed by the kubelet's
cadvisor, i.e. it no longer adds up metrics from pods. The host
level metrics are marked with the host_network label.

Note that on some infrastructures, an overlay network (the tunl0 network
device) is used for pod to pod communication. That network traffic is
shown separately on the chart. It is actually a subset of the inter node
traffic traffic that is shown for the eth0/ens.+ device.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>