Keep the container_network_.* metrics with an empty container label
The cadvisor metrics that are exposed by the kubelet
look different depending on which container runtime is used.

They can be accessed via:

  k proxy
  curl -s http://localhost:8001/api/v1/nodes/<node>/proxy/metrics/cadvisor

With docker, the container label's value is a dummy string "POD".
With containerd, the container label is empty.

Note that the network metrics are reported for a pod, not for a specific
container, because the containers of the pod share a network namespace
and the metrics are reported for a network namespace.

With the current configuration, the network metrics for containerd clusters
were dropped by Prometheus, because the container label's value was empty.

With this change, we keep these network metrics.

Note that the pipe symbol `|` is OR-ing both patterns:
- container_network.+;
  OR
- .+;.+

Note that we looked into whether the regular expression could be formatted
and commented for better readability. Unfortunately it seems that the
golang regex does not support the embedded comment syntax like java's
regular expression engine does. (See (?x) or Pattern.COMMENTS)
Furthermore, the line breaks can only be escaped in double quoted yaml
string so that they are not preserved or replaced with a space,
but then we can not add comments inside the string.
So we stick with the single line regular expressions for now.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>