Keep network metrics of the host
The network metrics are reported for the host, with the (cgroup) id "/".
They are also reported for each pod in the host network namespace that
we dropped in the previous commit.
With this change, we keep these host network metrics once for each host.

The name of the network interface device depends on the infrastructure:
- eth0 (classic)
- ens.+ (newer: see predictable interface naming)
+ tunl0 (overlay network for pod to pod communication; not on Azure)

Currently we observed this naming in the different infrastructures:

azure: eth0
ali:   ens5
os:    ens192
gcp:   ens4
aws:   ens5

Note that we need these metrics to show the network usage of a node,
for the Node details dashboard.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>