Use configured `syncPeriod` instead of hard-coded value
Previously, the `.controllers.seed.syncPeriod` value was defaulted to `30s`, however it was completely unused.
Now, it is defaulted to `10s` and used as part of the lifecycle controller. Earlier, the lifecycle controller was using the hard-coded `10s` value.

Also, the hard-coded `syncPeriodAfterExpiredMonitoringPeriod` value was dropped in favor of `syncPeriod` only (there is no clear benefit to use another sync period here).
Also, the duplicative read of the `Lease` object was dropped (there is no clear benefit to just read the very same Lease object twice in a row, especially since it's always done with the cached client).