Let kube-state-metrics (running in the garden ns) expose vpa metrics as well
There are 2 kinds of kube-state-metrics deployments in a seed:

- in each shoot control plane namespace
- in the garden namespace

The one running in each shoot control plane namespace is connected to the
shoot API server and hence exposes metrics about the shoot cluster.

The one running in the garden namespace is connected to the seed API server
and hence exposes metrics about the seed cluster.

Previously, only the shoot control plane kube-state-metrics deployment
had access to and exported the verticalpodautoscalers metrics.

With this change, the cluster role that is bound to the service account
that is used by the kube-state-metrics instance running in the garden
namespace is granted access to verticalpodautoscaler resources (get, list,
watch). Furthermore, the command line flags are extended so that vpa
metrics are exposed.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>