Allow list the vpa metrics for the "cache" prometheus in the garden ns
The kube-state-metrics in the garden namespace is scraped by the "cache"
prometheus. The newly exposed vpa metrics are allow listed with this
change.

The control plane prometheus is already scraping the kube-state-metrics
in the control plane namespace and allow listing some vpa metrics.
With this change, its allow list is adjusted so that now both prometheuses
keep the same set of vpa metrics:

- recommendation (target, upper bound, lower bound)
- min allowed, max allowed
- update mode

By having the same set of vpa metrics for both the seed and shoot clusters,
we can use the same promql queries in the dashboard for pods that are
running in the shoot cluster and in the control plane.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>