Federate the ksm metrics (cache prometheus -> aggregate prometheus)
Only federate the ksm metrics (including the new vpa metrics)
for the garden and extension namespaces.

The grafana deployment in the garden namespace uses the
aggregate-prometheus as a datasource, hence we need to federate these
ksm/vpa metrics to the aggregate-prometheus to be able to show them
on a dashboard.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>