Move extension roles test to e2e test suite
ClusterRole aggregration only works with a running kube-controller-manager,
thus it doesn't work in envtest.
One way to test it, would be to mimic the controller.
However, we have a proper setup for executing tests against a full-fledged
cluster (e2e tests), so rather use this one instead.
This makes sure, that the extension roles mechanism actually works e2e.