[GEP-20] Zone pinning for shoots on multi-zonal seeds 📌  (#6579)
* Add zone pinning label to seed namespaces

* Add zone information to seed namespace

After etcd is ready we assume that at least one pod with a volume has been scheduled. This is the earliest point in time when we can "pin" the zone, i.e. all remaining volumes also reside in that zone and for the future it needs to be assured that all pods (esp. the ones with volumes) are scheduled there.

* Add zone pinning for shoots on multi-zonal seeds

For non-HA or single-zonal HA shoot clusters, control-plane pods should be scheduled into a single zone only to avoid any cross zonal traffic.

* Address review comments

* Address review comments II

* Remove unneeded async assertions in integration tests

* Pass Gomega object to assertion func

* Fix losing zone information on reconciliation