Keep traffic from istio ingress gateway to the target pods in the same zone if possible. (#6653)
Supporting multiple availability zones has a number of benefits in terms of potential
availability. However, it also can have some downsides, especially with regards to
additional latency and potential cost increase. As availability zones have some spatial
distance between them there is some automatic latency added to traffic moving between
zones. In addition to that, there might be a cost involved in sending data between
zones.
These are the main reasons why it may be beneficial to keep traffic within an availability
zone if possible, i.e. if there are ready backend available within the zone. Even though
this might have positive impact with regards to reduced latency and costs, it is important
to note that this can also caused imbalance in terms of the request distribution, e.g.
backends in a zone without an istio ingress gateway might receive hardly any traffic as
long as the other backends can handle the load.
This change configures the istio ingress gateway to prioritize backends in the same zone
higher as those in different zones. In case no backend is available in the current zone
the request is routed to any of the other backends.
It is necessary to activate the outlier detection. Without it, istio will not honour the
fail-over priorities.
This change explicitly does not take care about scheduling workload in particular zones.