Switch `Seed` controller to `controller-runtime` (#6636)
* Update documentation

* Switch BackupBuckets check reconciler to native controller-runtime controller

* Switch Extensions check reconciler to native controller-runtime controller

* Integration test for secrets reconciler

* Switch secrets reconciler to native controller-runtime controller

* Use configured `syncPeriod` instead of hard-coded value

Previously, the `.controllers.seed.syncPeriod` value was defaulted to `30s`, however it was completely unused.
Now, it is defaulted to `10s` and used as part of the lifecycle controller. Earlier, the lifecycle controller was using the hard-coded `10s` value.

Also, the hard-coded `syncPeriodAfterExpiredMonitoringPeriod` value was dropped in favor of `syncPeriod` only (there is no clear benefit to use another sync period here).
Also, the duplicative read of the `Lease` object was dropped (there is no clear benefit to just read the very same Lease object twice in a row, especially since it's always done with the cached client).

* Integration test for lifecycle reconciler

* Switch lifecycle reconciler to native controller-runtime controller

* Drop no longer needed code

* Address PR review feedback

* Address PR review feedback

* Deflake ExtensionsCheck integration test suite

* Align BackupBucketsCheck integration test suite

* Fix invalid gardenlet Helm chart value config