Add a panel "Response Size Rate" to the API Server dashboard
This panel shows the rate of response sizes by resource and verb.

Outstandingly high values could give a hint for shoot operators on where
to start on the journey of reducing the load on the API servers,
by making sure to use WATCH requests instead of periodic LIST and
GET requests in the various controllers.

High values for WATCH requests could hint at reducing the rate of change
for the specific resource.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>