Introduce `manager.Runnable` for controlled bootstrapping
Typically, runnables added to the a manager are all started in parallel, hence uncontrolled. In the gardenlet case, we first want to perform a few bootstrap tasks (like validating the seed config or fetching a client certificate for the garden cluster) before we start the actual runnables (controllers, etc.).
The introduced `Runner` is capable of achieving this goal by first starting all bootstrap runnables and only adding the actual runnables once all bootstrap runnables completed successfully.