Add a panel "Response Size Rate" to the API Server dashboard (#6675)
* Export the API Server dashboard from Grafana (only generated changes)

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Fix the "API Server" variable

Previously it showed the job label, which is the same for each API server
instance.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Add a panel "Response Size Rate" to the API Server dashboard

This panel shows the rate of response sizes by resource and verb.

Outstandingly high values could give a hint for shoot operators on where
to start on the journey of reducing the load on the API servers,
by making sure to use WATCH requests instead of periodic LIST and
GET requests in the various controllers.

High values for WATCH requests could hint at reducing the rate of change
for the specific resource.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Minor: use the apiserver variable instead of a hard coded value

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>