Watch `ManagedResource`s and map to `ControllerInstallation`s
The main `ControllerInstallation` controller now adds a label to the created `ManagedResource` in the seed so that the care controller can map the `ManagedResource` to the owning `ControllerInstallation`.
Also, the predicate about changed conditions from the seed's `ExtensionsCheck` controller was moved to the general predicate utils so that it can be reused in the care controller now.