Introduce `controller-runtime` manager in `gardenlet` (#6688)
* Drop useless/duplicate tests

These tests don't belong to the `controller` packages but rather to the `imagevector` package.

* Streamline `Log{Level,Format}` types in `GardenletConfiguration`

Pointer types are removed now just like in the component configs for GCM, scheduler, GAC, etc.

* Rework `GardenletConfiguration` server settings

The config structure is similar to admission-controller and controller-manager now and fits to the `manager.Options` structure as well.

* Move `AddAllFieldIndexes` to `api/indexer` package

So that it can be reused in `gardenlet` later on

* Log expiration date of client certificate in seed controller

Previously, this was logged in the gardenlet's `app.go`, but the seed controller is the only entity which requires the client certificate object. So, better move this here instead of transporting the information through many different intermediate instances.

* Move gardenlet identity creation to `controller` package

* Make the `healthz.Manager` implement `manager.Runnable`

The health manager in gardenlet will be registered as `manager.Runnable` in the future.

* Make the `clientmap.ClientMap` implement `manager.Runnable`

The client map for shoots in gardenlet will be registered as `manager.Runnable` in the future.

* Make the seed config check implement `manager.Runnable`

* Make garden kubeconfig bootstrapping implement `manager.Runnable`

* Make `certificatemanager.ScheduleCertificateRotation` block

This function will be ran as `manager.Runnable`, so make sure that it blocks as long as it is actively waiting (previously the function was starting a goroutine and returning immediately, but this doesn't play nicely with `manager.Runnable` way of running code).

* Adapt certificate rotation coding

So that garden client is no longer fetched from client map

* Simplify shoot client map

The gardenlet will have the garden and seed clients already available when creating the client map, so we can pass them right away.

* Introduce `manager.Runnable` for controlled bootstrapping

Typically, runnables added to the a manager are all started in parallel, hence uncontrolled. In the gardenlet case, we first want to perform a few bootstrap tasks (like validating the seed config or fetching a client certificate for the garden cluster) before we start the actual runnables (controllers, etc.).
The introduced `Runner` is capable of achieving this goal by first starting all bootstrap runnables and only adding the actual runnables once all bootstrap runnables completed successfully.

* Make `kubernetes.FallbackClient` public

Earlier, this client was used when instantiating a new `kubernetes.Interface`, however the gardenlet will now instantiate the clients using the controller-runtime's manager. Hence, we have to make this function public so that we can use provide it to the `manager.Options.NewClientFunc`.

* Allow to provide existing cache, reader and client when creating `kubernetes.Interface`

The gardenlet will have these entities already available when it creates a `kubernetes.Interface` for the seed, so no need to create a duplicate cache, client, etc.

* Rework gardenlet options similar to other components

* Rework gardenlet main entrypoint

* Drop custom HTTP(S) server package

We will be only using the HTTP servers that already come with the controller-runtime's `Manager`, so we don't need this package anymore.

* Drop no longer required code in client maps

We don't need the garden or seed client maps anymore, and we allso don't need the delegating client map. Only shoot client maps will remain.

* Adapt gardenlet's legacy controller factory

* Adapt `operation` builder

* Adapt gardenlet controller interfaces

The will be provided with the `cluster.Cluster` or client objects from the factory when the controllers are instantiated.
Also, there will be no ClientMap anymore for garden/seed, so related coding can be cleaned up/deleted.

* Rename `K8sSeedClient` to `SeedClientSet` in `Operation`

This is to have aligned names of the clients across code and components. Also, "ClientSet` hints that it's not `client.Client` but `kubernetes.Interface`.

* Rename `K8sShootClient` to `ShootClientSet` in `Operation`

This is to have aligned names of the clients across code and components. Also, "ClientSet` hints that it's not `client.Client` but `kubernetes.Interface`.

* Rename `ClientMap` to `ShootClientMap` in `Operation`

In the future, there will be no client maps for garden or seed clusters anymore (no longer needed). Hence, let's make sure that this client map is for shoot clusters only.

* Rename `K8sGardenClient` to `GardenClient` in `Operation`

This is to have aligned names of the clients across code and components. Also, "Client` hints that it's `client.Client` and not `kubernetes.Interface`.

* Drop `botanist.InitializeSeedClient` method

The gardenlet will have the seed client always available, no need to initialize it anymore.

* Address PR review feedback

* Address PR review feedback