Watch objects in health controller
Start additional watches for all GVKs encountered in `ManagedResource.status.resources`.
Objects are mapped to the owning ManagedResource to trigger the health controller as soon
as a change to the health status of any managed object happens.

Fetching events is split from health.CheckService so that we can call CheckHealth
without fetching events in the healthStatusChanged predicate. In the predicate,
we are not interested in the details but only in changes to the health status.
Events are now fetched separately after a failed health check in the reconciler.