Remove vpa-exporter
The vpa-exporter is part of the vpa component.

This commit removes all the artifacts related to
the vpa-exporter, including

- monitoring configuration
- unit tests

It also adjusts

- the prometheus network policy
- the priority class documentation

Note that the resource manager will remove the existing vpa-exporter
artifacts from the seed cluster so explicit deletion is not needed,
it is sufficient to just remove these resources.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>