Add a script to export the state of the cluster relevant for this PR
We can use the (remote) local setup and deploy the base commit of this PR,
and export the relevant state of the cluster with this script.

Then, we can upgrade to this PR and restart the gardenlet and let the
changes be applied and export the relevant state again.

The diff shows the changes that are triggered by the PR. The resulting
diff might be helpful when reviewing the PR.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>