Enhance resource-manager health controller (#6770)
* Watch objects in health controller

Start additional watches for all GVKs encountered in `ManagedResource.status.resources`.
Objects are mapped to the owning ManagedResource to trigger the health controller as soon
as a change to the health status of any managed object happens.

Fetching events is split from health.CheckService so that we can call CheckHealth
without fetching events in the healthStatusChanged predicate. In the predicate,
we are not interested in the details but only in changes to the health status.
Events are now fetched separately after a failed health check in the reconciler.

* Use target cache instead of source cache

* Make health controller logs less noisy

* Eliminate unstructured objects from health controller

We now either use typed objects or metadata-only objects (for GVKs not registered in target scheme).
With this, we can fully leverage the cache of the target cluster (if enabled), even for objects we
don't know. For them, we only care about their presence, so metadata-only is enough.

* Nits

* Remove confusing early exit for metadata-only events

CheckHealth can now properly handle metadata-only objects, no need
to treat them specifically in the predicate.

* Go back to map+RWMutex, properly store after successfull add