Support running "make generate" in the (dev) "remote-local-setup" (#6780)
Gardener developers need to execute `make generate` if the source code
changes require adjusting the versioned generated code.

This step is very CPU intensive and somewhat time consuming (5+ minutes).

This commit adjusts the development version of the remote-local-setup
so that `make generate` can be executed in the remote pod in the cloud,
saving some CPU cycles for the local machine.

The gcompat package is the GNU C Library compatibility layer for musl,
which is needed for this alpine based docker image.

`parallel` and `tar` are needed by `make generate`. The simpler BusyBox
version of `tar` is lacking some features.

`strace` is generally useful for debugging purposes.

code generation currently needs the old style Go setup: hence
`GOPATH` is set to `~/go` and the gardener git repository is symlinked
to `~/go/src/github.com/gardener/gardener`.

The repository is symlinked only: the default working directory is
still `~/gardener`, it is preferred for its shorter path.

The symbolic link
  ln -s ~/gardener/hack/tools/bin/protoc-gen-gogo ~/gardener/protoc-gen-gogo
is needed when executing `make generate`: `protoc` tries to execute
`protoc-gen-gogo` in the current directory which is the repository root,
and does not seem to search the `PATH`. (`strace -f` was useful to figure
that out).
Hence a symbolic link for `protoc-gen-gogo` is added to the repository
root.

Note that I'm not sure why this is not an issue when `make generate`
is executed on the laptop but I did not attempt to debug that on the Mac.

Note also that the 8GB memory limit was not sufficient for the parallel
code generation: a few Go compile processes were OOM killed, so the
memory requests and limits are doubled now.

With this change `make generate` can be executed in the
`~/go/src/github.com/gardener/gardener` folder and takes about 5 minutes
with 8 cores in the cloud:

```
cd ~/go/src/github.com/gardener/gardener
make generate
```