[gardenlet] Switch `ControllerInstallation` controller to `controller-runtime` (#6733)
* Use `spec.seedRef.name` field selector for `ControllerInstallation` cache

This way, we don't have to filter unrelated `ControllerInstallation`s at
the controller level (e.g., with predicates). The underlying
controller-runtime cache already makes sure that we only see objects
which are related to our seed now.

* Properly register indexed fields in registry

Without this, the field selector watch for the gardenlet does not work.

* -------- empty separator commit --------

* Add documentation for care reconciler

* Switch care reconciler to native controller-runtime controller

* Limit exponential backoff to sync period

Similar to other refactored care controllers.
Now we can properly/cleanly handle and return errors occuring during reconciliation.

* Watch `ManagedResource`s and map to `ControllerInstallation`s

The main `ControllerInstallation` controller now adds a label to the created `ManagedResource` in the seed so that the care controller can map the `ManagedResource` to the owning `ControllerInstallation`.
Also, the predicate about changed conditions from the seed's `ExtensionsCheck` controller was moved to the general predicate utils so that it can be reused in the care controller now.

* -------- empty separator commit --------

* Add documentation for main reconciler

* Add integration test for main `ControllerInstallation` controller

* Switch main reconciler to native controller-runtime controller

* Replace `Reconciler.isResponsible` function with a predicate

* Drop no longer required code

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback