Added checks for non-gardener managed seed clusters to ensure correct seed configurations.
Clusters managed by gardener embed information about the node, pod and service networks
inside the cluster as a config map. Other kubernetes clusters do not offer this information
in a standard way.
To ensure that the seed configuration fits to the reality of the seed cluster this change
introduces a heuristic check that the configured node network is valid for the nodes seen
in the cluster. The same applies to the pods and services.
(The amount of entities being checked is limited so that large clusters do not delay the
gardenlet start-up too much.)