Merge the dashboards that are contributed by the extensions
According to the monitoring extension contract, extensions can
contribute dashboards for the user and the operator Grafana.

We do not change the extension contract here, to avoid adjusting all
the extensions.

The implicit behavioral change is that dashboard that were contributed either to
the user Grafana or to the operator Grafana, will all end up in the single new
user Grafana instance.

Some extensions contribute the same dashboard to both instances, although the
operator Grafana already was a superset of the user Grafana. These duplicates
are implicitly removed at a later point.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>