Watch `ManagedResource`s and map to `Seed`
The main `Seed` controller now adds a label to the created `ManagedResource` in the seed so that the care controller can map the `ManagedResource` to the `Seed`.
Also, the predicate about changed conditions from the `ManagedResource` controller was moved to the general predicate utils so that it can be reused in the various care controllers now.
In addition, the `ControllerInstallation` care controller was made "more precise" by only caring about `ManagedResource`s in the `garden` namespace.