[gardenlet] Switch `Seed` controller to `controller-runtime` (#6836)
* Add documentation for lease reconciler

* Add integration test for lease reconciler

* Switch lease reconciler to native controller-runtime controller

* Fix flaky unit test of lease reconciler

Earlier:

```
3660 runs so far, 1 failures
```

Now:

```
37412 runs so far, 0 failures
```

* -------- empty separator commit --------

* Add documentation for care reconciler

* Add integration test for care reconciler

* Switch care reconciler to native controller-runtime controller

* Watch `ManagedResource`s and map to `Seed`

The main `Seed` controller now adds a label to the created `ManagedResource` in the seed so that the care controller can map the `ManagedResource` to the `Seed`.
Also, the predicate about changed conditions from the `ManagedResource` controller was moved to the general predicate utils so that it can be reused in the various care controllers now.
In addition, the `ControllerInstallation` care controller was made "more precise" by only caring about `ManagedResource`s in the `garden` namespace.

* -------- empty separator commit --------

* Add documentation for main reconciler

* Add integration test for main reconciler

This test is far from complete (there are way more assertions one can
make) since this reconciler is doing quite a lot of things. However,
it's a good basis and serves well as integration test for the
controller. It can be extended in the future with more cases.

* Switch main reconciler to native controller-runtime controller

* Extract seed deletion logic into dedicated function

Also, move related logic from `pkg/operation/seed` to the controller package so that it's not unnecessarily scattered and distributed.

* Extract seed reconciliation logic into dedicated function

Also, move related logic from `pkg/operation/seed` to the controller package so that it's not unnecessarily scattered and distributed.

* De-duplicate assignment of seed cluster identity

* -------- empty separator commit --------

* Drop useless `concurrentSyncs` field in `SeedControllerConfiguration`

The gardenlet is only responsible for exactly one Seed, hence it doesn't make sense to configure more than one concurrent sync.

* Drop no longer needed code

* Do not determine identity twice

Otherwise, the gardenlet will generate two identities in case it is not running as a pod since it generates random strings as part of `determineIdentity`.
Once legacy shoot controller is refactored we can safely move it to the `pkg/gardenlet/controllers` package again.

* Do not generally restrict gardenlet's `Seed` cache

In some cases (for CPM, for example), the gardenlet has to read other `Seed` resources. Without this commit, they aren't available in the cache and the gardenlet does not fallback to making real API requests, hence it fails.

* Address PR review feedback

* Address PR review feedback