Mitigation for a cgroup issue with the shoot node in the local setup
The cadvisor metrics are reported twice for the pod that represents the shoot
node in the local setup. It seems that the cgroups of that "node" pod are mixed
up: metrics are reported by the cadisor of the KinD node with a CGROUP_PATH
cgroup path and also with a repeated cgroup path: CGROUP_PATH/CGROUP_PATH. The
metrics with CGROUP_PATH contain the correct value, the metrics with
CGROUP_PATH/CGROUP_PATH always return 0. (There is a CGROUP_PATH/CGROUP_PATH
cgroup path on the KinD node but it does not contain any tasks, and hence the
memory or CPU usage of that cgroup is 0.)

With the current Prometheus scrape configuration the 2 metrics are mapped to the
same time series because the id label is dropped to reduce the cardinality of
the metrics. Hence the resulting time series contains randomly either the
correct reading or 0. This causes some confusion on the metering Grafana
dashboards in the local setup.

The mitigation with this commit is to detect the repeated path
CGROUP_PATH/CGROUP_PATH specific to the local setup and discard the scraped
values with the repeated path: CGROUP_PATH/CGROUP_PATH.

The regular expression is specific enough (contains the repetition and the term
docker in the middle) so that in the regular, non "local setup", no cadvisor
metrics are expected to match and hence this change is a no-op for the "real"
setup.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>