Do not wipe the "cache" Prometheus' disk on restart
This commit reverts
https://github.com/gardener/gardener/pull/884/commits/86b8203fbf94a4e2fda53e1b11a4d15068f635be
86b8203fb Delete WAL when prometheus starts (2019-03-29)

Back in 2019 the "cache" Prometheus was configured to delete its state each time
it was (re)started, due to issues like "corrupt WAL file, full disk".

These issues are no longer prevalent so this stateless configuration is changed
back to the default, "stateful" configuration, so that the "cache" Prometheus
can have a persistent state that is sustained beyond restarts.

This is necessary to execute recording rules that are self referential and
access an older value of the same series, an idiom when accumulating values like
this:

  - record: cpu_usage:sum_over_time
    expr: |2
        cpu_usage
      +
        (
            last_over_time(cpu_usage:sum_over_time[10m])
          or
            cpu_usage * 0
        )

Here, the new value of `cpu_usage:sum_over_time` depends on an older value of
the same series: `cpu_usage:sum_over_time`. Without persistent state, these self
referential recording rules would "restart" e.g. the accumulation when Prometheus
loses its state due to a restart.

Note that the "cache" Prometheus still has a fairly short retention, which was
slightly increased from 2 hours to 1 day, from 1GB to 5GB to make better use of
the 10GB disk which is currently provisioned for the cache Prometheus.

Its main role is to scrape the cluster wide cadvisor and kube-state-metrics
metrics so that the numerous control plane Prometheus instances can federate the
relevant subset of these metrics. In addition, it can now execute recording
rules, even ones that rely on persistent state, to calculate and aggregate
metrics for federation. An example for these recording rules are the new
metering calculations that are added in this PR.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>