Use service instead of endpoint "service discovery" for federation
We are only federating stateless metrics from the "cache" Prometheus, i.e.
metrics that are computed only based on the current values of some other
metrics. Stateful metrics (like :sum_over_time) are not federated from the
"cache" Prometheus.

This allows to later use a "higher availability" pair of "cache" Prometheus
instances and federate an equivalent metric from either of the 2 instances.

In that case the identity of the "cache" Prometheus instance does not matter: so
instead of using endpoint based service discovery, which saves the pod IP into
the instance label and thereby creates new series when the pod IP changes, we
use service based service discovery which saves the name of the service in the
instance label. The name of the service is stable, it does not change when the
pod IP changes.

Note that the namespace is filter in the service discovery and hence Prometheus
need not watch all the services in all the namespaces: only the services in the
garden namespace.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>