Calculate the monthly average resource usage of the shoots (!)
The monthly average resource (CPU, Memory, Network and Disk) usage and requests
of the shoot control planes that are hosted in this seed are calculated by an
elaborate set of "stateful" recording rules in the "cache" Prometheus. This
calculation should be repeated in federating Prometheus instances instead of
federating the results of these recording rules. That way we could have 2
"cache" Prometheus instances which are federated via a k8s service, i.e. a
random instance is federated, for higher availability and reduced network costs.

These recording rules are an optimization to provide an alternative to the
otherwise correct but inefficient / expensive PromQL queries that calculate the
monthly average resource usage with a range vector selector and an _over_time
aggregation function.

  avg_over_time(metering:cpu_usage:sum_by_namespace[30d])

These recording rules allow to show the precomputed monthly average usage values
on a Grafana dashboard in an interactive way.

If a shoot cluster is hibernated, it does not consume CPU, Memory or Network,
but it still consumes Disk space. So the `:memory_usage_seconds` and the
`:disk_usage_seconds` can be used to measure the lifetime of the control plane
and also the duration when the cluster was hibernated. Note that we picked the
memory usage to measure the lifetime of the control plane (instead of CPU or
Network) because if a process is running, it necessarily consumes memory, but it
might not consume any CPU or network. Note that when a cluster is created, the
creation process can start with some pods of a deployment (without persistent
volumes) before creating pods of a statefulset (with persistent volumes). So the
time when memory is used might start a few minutes earlier than the time when
persistent volumes are used. Furthermore, during the teardown, it might take
some minutes to delete the persistent volume claims after all the pods have been
removed. This means that even if a cluster is never hibernated, the 2 durations
might differ a bit.

The average is calculated by accumulating the momentary usage values in the
`:sum_over_time series` and dividing them by either the `:memory_usage_seconds`
(for Memory, CPU and Network) or `:disk_usage_seconds` (for Disk).

- X :sum_over_time
- X :avg_over_time

When a shoot cluster is deleted and all its resources (deployments,
statefulsets, pods, volumes and the namespace) are cleaned up, the raw metrics
are not going to be emitted anymore and hence the respective derived metrics
will also become absent. This is the expected behaviour with time series in
Prometheus and helps to reduce the load due to churn: many short lived clusters.

If we want to list all the shoot clusters that existed during a month, including
also the shoot clusters that existed only for a few hours, we'll need to use a
range vector in combination with an _over_time function, e.g.
`last_over_time(...[30d])`. This query is fairly expensive because Prometheus
would have to go through all the series (could be 10k+) for the 43k scrapes
during a month. Trading in disk usage to reduce the CPU and memory usage of
those queries, the absent series (average and ratio) are continued for the
remaining time in the month, and hence regular instant vector queries can be
used to list all the shoot clusters that existed in a month.

- X :this_month

The 8 calculated metrics in combination with the 3 aspects above yield 24 new
series and hence a small shell script based template is used to generate the
respective recording rules.

Note that by adding the date (year, month) label to the time series, the sum and
average calculations are naturally reset at the beginning of the month. The
:this_month "retention" is also reset similarly at the beginning of the month.

The `last_over_time(metering:...:sum_over_time[10m])` is used to increase the
default 5 min look back time when querying the current value of a time series to
10 minutes. This should safely bridge the downtime due to rolling the nodes and
restarting the prometheus instance.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>