Remove `seedTemplate` field from ManagedSeed API (#6972)
* Remove SeedTemplate field from ManagedSeed

* Remove more seedTemplate usage

* Adjust SeedRestriction webhook

Some cases are obsolete with the deprecation of `seedTemplate`.
The only sophisticated use-case for the webhook to check is when the
"parent gardenlet" deletes the seed object in the scope of a ManagedSeed
object deletion.

* Drop `connect` case from seed-restriction webhook

Checking or explicitly denying the `connect` operation is redundant
since this the operation will never be sent by the Kube-Apiserver
if not configured by the `validatingwebhookconfiguration`.

* Add protobuf number test

* Address review comments