Fix a "many-to-many matching" error case in a recording rule (#7097)
In 826af63e8e5 Calculate the monthly average resource usage of the shoot control planes

recording rules were added to calculate the resource usage of the shoot control
planes and we noticed that the recording rule which calculates the persistent
disk usage by owner can throw an error:

  metering:persistent_volume_usage:sum_by_namespace_owner:
    ...
    many-to-many matching not allowed: matching labels must be unique on one side

when a statfulset's pod with a persistent volume is rescheduled to a different
node.

During that transition it can happen that the kubelet metrics about the disk
usage are reported (or at least returned by the PromQL queries) for both the old
and the new nodes. This means that for a single persistent volume we get 2 usage
metrics instead of 1 and hence the many-to-many vector matching error is
reported during this transition period, typically for a single evaluation
iteration.

To avoid this temporary error condition we can take the maximum of the duplicate
metrics, i.e. the maximum for each conceptually unique namespace, persistent
volume claim combination. By taking the maximum, we make sure to have only a
single time series for any namespace, persistent volume claim combination and
hence we can avoid the many-to-many matching error above.

The same underlying behavior could also impact the calculation of the
metering:persistent_volume_usage:sum_by_namespace recording rule by adding up
the duplicated disk usage metrics and hence double counting the disk usage of a
single disk. By taking the maximum first, we can avoid that temporary glitch.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>