Adapt cluster-proportional autoscaling of coredns to new high-availability handling.
With PR #6989 high-availability handling was introduced to coredns and other components.
While it works nicely with the default scenario with horizontal pod autoscaler, it fails
in combination with the cluster-proportional autoscaling approach. In the latter case
the autoscaler and the gardener-resource-manager race towards changing the replica count
of the coredns deployment. As a result, pods are created/terminated constantly.
With this change, the replica mutation of the high-availability handling is disabled in
case coredns uses the cluster-proportional autoscaling approach.