Extend seed api load balancer settings for zone-specific traffic handling.
To be able to reduce cross-zonal traffic in a multi-zonal seed, the seed specification
was extended to include load balancer settings per availability zone. The zone settings
in the seed specification allow operators to specify annotations, which will only affect
load balancers for a specific zone. These annotations can be used, depending on the
infrastructure specific cloud controller implementation, to for example bind a load
balancer to a specific subnet or availability zone and hence prevent cross-zonal traffic
between the cloud infrastructure load balancer and the corresponding istio ingress
gateways.
In addition to that, the external traffic policy of the istio ingress gateways is now
configurable in the seed specification as well. This setting is valid for the default
istio ingress gateway as well as the zone-specific istio ingress gateways. It will
most likely be set consistently for all, but operators can also choose to handle it
specific for each zone.
The default external traffic policy is cluster meaning that all traffic from the load
balancer will be sent to any cluster node, which then itself will redirect the traffic
to the actual receiving pod. This approach adds a node to the data path and may cross
the zone boundaries twice.
Using external traffic policy local drops the additional node, i.e. only cluster nodes
with corresponding backend pods will be in the list of backends of the load balancer.
However, this has multiple implications. The health check port in this scenario is
exposed by kube-proxy, i.e. if kube-proxy is not working on a node a corresponding
pod on the node will not receive traffic from the load balancer as the load balancer
will see a failing health check. (This is quite different from ordinary service
routing where kube-proxy is only responsible for setup, but does not need to run for
its operation.) Furthermore, load balancing may become imbalanced if multiple pods run
on the same node because load balancers will split the load equally among the nodes
and not among the pods.
The aspect of kube-proxy being responsible for the health check is accepted. The second
issue is mitigated with node anti affinities.
As a side effect/refactoring of the code, the exposure class handler has been removed
from pkg/operation/seed|shoot. The new unified accessor functions make storing the
exposure class name/handler unnecessary. Removing the code simplifies the code base.