Create one istio ingress gateway per availability zone to reduce cross-zonal traffic.
To reduce cross-zonal traffic and ensure that traffic stays within an availability zone
for traffic touching the istio ingress gateway, new istio ingress gateways are created.
If the number of zones as specified in seed.spec.provider.zones is larger than one an
additional istio ingress gateway is created per availability zone. The default istio
ingress gateway acts as the multi-zonal entry point.
The same holds for exposure classes, i.e. the existing istio ingress gateway for the
exposure class acts as the multi-zonal entry point and additional istio ingress gateways
are created for each availability zone.
This allows single-zone shoot control planes to be exposed via a single-zone istio
ingress gateway in the same availability zone. Multi-zone shoot control planes are
exposed via the default istio ingress gateway. (This works analogously for exposure
classes.)
The existing cleanup logic for orphaned exposure classes has been adapted to also
handle the zone-specific istio ingress gateways.
The namespaces of the zone specific istio ingress gateways use the same name as the
corresponding default gateway, but have '--' and the zone name appended to it. As
exposure classes allowed the exposure class name to cover all available space, the
maximum exposure class name length had to be restricted. In case the name gets to long
a hash of the zone is used instead.
Due to the fact that it is not possible to simply change the selector of an existing
deployment, the labels of the zone-specific istio ingress gateways use label key, which
already exist for the default gateway or exposure classes. For non-exposure classes
zone-specific istio ingress gateways 'istio' is used as the label key and the label
value will be 'ingressgateway--zone--' and the zone name. For exposure classes,
'gardener.cloud/role' is the label key and 'exposureclass-handler--zone--' + zone name
is the label value.
As there may already be existing single-zone shoot cluster control planes before this
change is introduced, these control planes will have their external endpoint migrated
to the corresponding single-zone istio ingress gateway. This implies a change in the
externally visible IP address of the kube-apiserver. In case this address was used in
IP allow lists they have to be updated.